# src/seed/seed.controller.ts
import { Controller, Get } from '@nestjs/common';
import { SeedService } from './seed.service';
import { Auth } from '../auth/decorators';
import { ValidRoles } from '../auth/interfaces';
import { ApiBearerAuth, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';

@ApiTags("Seed")
@Controller('seed')
export class SeedController {
  constructor(private readonly seedService: SeedService) {}

  @Get()
  @Auth(ValidRoles.admin)
  @ApiOperation({
    summary: 'Execute database seed (admin only)',
    description:
      'Deletes all existing data and inserts initial users and products',
  })
  @ApiBearerAuth('access-token')
  @ApiResponse({
    status: 200,
    description: 'Seed executed successfully',
    schema: {
      example: 'SEED EXCECUTED',
    },
  })
  @ApiResponse({
    status: 401,
    description: 'Unauthorized',
  })
  @ApiResponse({
    status: 403,
    description: 'Forbidden – admin role required',
  })
  excecuteSeed() {
    return this.seedService.runSeed();
  }

}