# swagger.json
{"openapi":"3.0.0","paths":{"/api/products":{"post":{"operationId":"ProductsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":"Product was created succesfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidenn. Token related"}},"tags":["Products"]},"get":{"operationId":"ProductsController_findAll","parameters":[{"name":"limit","required":false,"in":"query","description":"Number of items to return","schema":{"default":10,"example":10,"type":"number"}},{"name":"offset","required":false,"in":"query","description":"Number of items to skip","schema":{"default":0,"example":0,"type":"number"}}],"responses":{"200":{"description":"List of products","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}},"summary":"Get all products (paginated)","tags":["Products"]}},"/api/products/{term}":{"get":{"operationId":"ProductsController_findOne","parameters":[{"name":"term","required":true,"in":"path","description":"Product UUID, slug or title","schema":{"example":"tesla_t_shirt","type":"string"}}],"responses":{"200":{"description":"Product found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"404":{"description":"Product not found"},"500":{"description":"Unexpected server error"}},"summary":"Get a product by id, slug or title","tags":["Products"]}},"/api/products/{id}":{"patch":{"operationId":"ProductsController_update","parameters":[{"name":"id","required":true,"in":"path","description":"Product UUID","schema":{"example":"1fab36d4-0552-41d0-a978-338dc452014e","type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":"Product updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"400":{"description":"Bad request (validation or duplicated field)"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden – admin role required"},"404":{"description":"Product not found"}},"security":[{"access-token":[]}],"summary":"Update a product (admin only)","tags":["Products"]},"delete":{"operationId":"ProductsController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"Product UUID","schema":{"example":"1fab36d4-0552-41d0-a978-338dc452014e","type":"string"}}],"responses":{"200":{"description":"Product deleted successfully","content":{"application/json":{"schema":{"example":"Producto eliminado correctamente"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden – admin role required"},"404":{"description":"Product not found"}},"security":[{"access-token":[]}],"summary":"Delete a product (admin only)","tags":["Products"]}},"/api/auth/register":{"post":{"operationId":"AuthController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User registered successfully","content":{"application/json":{"schema":{"example":{"id":"c1b2f9c4-0c55-4c9a-9e8d-4e4c5f7e1234","email":"user@email.com","fullName":"John Doe","isActive":true,"roles":["user"],"token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}}}}},"400":{"description":"Bad request (email already exists)"},"500":{"description":"Internal server error"}},"summary":"Register a new user","tags":["Auth"]}},"/api/auth/login":{"post":{"operationId":"AuthController_loginUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":"User logged in successfully","content":{"application/json":{"schema":{"example":{"id":"c1b2f9c4-0c55-4c9a-9e8d-4e4c5f7e1234","email":"user@email.com","token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}}}}},"401":{"description":"Invalid credentials"}},"summary":"Login user","tags":["Auth"]}},"/api/auth/check-status":{"get":{"operationId":"AuthController_checkAuthStatus","parameters":[],"responses":{"200":{"description":"Valid token","content":{"application/json":{"schema":{"example":{"id":"c1b2f9c4-0c55-4c9a-9e8d-4e4c5f7e1234","email":"user@email.com","fullName":"John Doe","roles":["user"],"token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}}}}},"401":{"description":"Unauthorized"}},"security":[{"access-token":[]}],"summary":"Check authentication status and renew token","tags":["Auth"]}},"/api/auth/private":{"get":{"operationId":"AuthController_testingPrivateRoute","parameters":[],"responses":{"200":{"description":"Access granted","content":{"application/json":{"schema":{"example":{"ok":true,"user":{"id":"c1b2f9c4-0c55-4c9a-9e8d-4e4c5f7e1234","email":"admin@email.com","roles":["admin"]}}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden – insufficient role"}},"security":[{"access-token":[]}],"summary":"Test private route (admin / superUser only)","tags":["Auth"]}},"/api/seed":{"get":{"description":"Deletes all existing data and inserts initial users and products","operationId":"SeedController_excecuteSeed","parameters":[],"responses":{"200":{"description":"Seed executed successfully","content":{"application/json":{"schema":{"example":"SEED EXCECUTED"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden – admin role required"}},"security":[{"access-token":[]}],"summary":"Execute database seed (admin only)","tags":["Seed"]}},"/api/files/product/{imageName}":{"get":{"operationId":"FilesController_findProductImage","parameters":[{"name":"imageName","required":true,"in":"path","description":"Product image filename","schema":{"example":"product-123-main.jpg","type":"string"}}],"responses":{"200":{"description":"Image file"},"400":{"description":"Image not found"}},"summary":"Get a product image by filename","tags":["Files - Get and Upload"]}},"/api/files/product":{"post":{"operationId":"FilesController_uploadFile","parameters":[],"requestBody":{"required":true,"description":"Product image file","content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":"Image uploaded successfully","content":{"application/json":{"schema":{"example":{"secureUrl":"http://localhost:3000/api/files/product/product-123-main.jpg"}}}}},"400":{"description":"Invalid file type (only images allowed)"}},"summary":"Upload a product image","tags":["Files - Get and Upload"]}}},"info":{"title":"Teslo Nest Api","description":"Teslo shop endpoints","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateProductDto":{"type":"object","properties":{"title":{"type":"string","description":"Product title (unique)","nullable":false,"minLength":1},"price":{"type":"number","description":"Product price","example":49.99,"minimum":0},"description":{"type":"string","description":"Detailed product description","example":"Premium cotton t-shirt with relaxed fit"},"slug":{"type":"string","description":"SEO friendly slug (auto-generated if not provided)","example":"tesla_t_shirt"},"stock":{"type":"number","description":"Available product stock","example":20},"sizes":{"description":"Available product sizes","example":["S","M","L","XL"],"type":"array","items":{"type":"string"}},"gender":{"type":"string","description":"Product gender category","example":"unisex","enum":["men","woman","kid","unisex"]},"tags":{"description":"Product tags","example":["shirt","casual","summer"],"type":"array","items":{"type":"string"}},"images":{"description":"Product image filenames or URLs","example":["img1.jpg","img2.jpg"],"type":"array","items":{"type":"string"}}},"required":["title","sizes","gender"]},"Product":{"type":"object","properties":{"id":{"type":"string","example":"1fab36d4-0552-41d0-a978-338dc452014e","description":"Product ID","uniqueItems":true,"format":"uuid"},"title":{"type":"string","example":"T-Shirt","description":"Product Title","uniqueItems":true},"price":{"type":"number","example":0,"description":"Product price","minimum":0},"description":{"type":"string","example":"Introducing the Tesla Raven Collection. The Women's Raven Joggers have a premium, relaxed silhouette made from a sustainable bamboo cotton blend. The joggers feature a subtle thermoplastic polyurethane Tesla wordmark and T logo and a french terry interior for a cozy look and feel in every season. Pair them with your Raven Slouchy Crew Sweatshirt, Raven Lightweight Zip Up Jacket or other favorite on the go fit. Made from 70% bamboo and 30% cotton.","description":"Product description"},"slug":{"type":"string","example":"t_shirt_teslo","description":"Product SLUG - for SEO","uniqueItems":true},"stock":{"type":"number","example":10,"description":"Product stock","default":0},"sizes":{"example":["M","XL","XXL"],"description":"Product sizes","type":"array","items":{"type":"string"}},"gender":{"type":"string","example":"women","description":"Product gender","enum":["men","woman","kid","unisex"]},"tags":{"description":"Product tags","example":["shirt","casual","summer"],"type":"array","items":{"type":"string"}},"images":{"description":"Product images URLs","example":["img1.jpg","img2.jpg"],"type":"array","items":{"type":"string"}},"user":{"description":"User who created or last updated the product","allOf":[{"$ref":"#/components/schemas/User"}]}},"required":["id","title","price","description","slug","stock","sizes","gender","tags","images","user"]},"User":{"type":"object","properties":{"id":{"type":"string","description":"User unique identifier","example":"c1b2f9c4-0c55-4c9a-9e8d-4e4c5f7e1234","format":"uuid"},"email":{"type":"string","description":"User email address (unique)","example":"user@email.com"},"password":{"type":"string","description":"User password (never returned in responses)","writeOnly":true,"example":"Password123!"},"fullName":{"type":"string","description":"User full name","example":"John Doe"},"isActive":{"type":"boolean","description":"Indicates if the user is active","example":true},"roles":{"description":"User roles","example":["admin","user"],"type":"array","items":{"type":"string"}},"product":{"description":"Products created by the user","type":"array","items":{"$ref":"#/components/schemas/Product"}}},"required":["id","email","password","fullName","isActive","roles","product"]},"UpdateProductDto":{"type":"object","properties":{"title":{"type":"string","description":"Product title (unique)","nullable":false,"minLength":1},"price":{"type":"number","description":"Product price","example":49.99,"minimum":0},"description":{"type":"string","description":"Detailed product description","example":"Premium cotton t-shirt with relaxed fit"},"slug":{"type":"string","description":"SEO friendly slug (auto-generated if not provided)","example":"tesla_t_shirt"},"stock":{"type":"number","description":"Available product stock","example":20},"sizes":{"description":"Available product sizes","example":["S","M","L","XL"],"type":"array","items":{"type":"string"}},"gender":{"type":"string","description":"Product gender category","example":"unisex","enum":["men","woman","kid","unisex"]},"tags":{"description":"Product tags","example":["shirt","casual","summer"],"type":"array","items":{"type":"string"}},"images":{"description":"Product image filenames or URLs","example":["img1.jpg","img2.jpg"],"type":"array","items":{"type":"string"}}}},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","description":"User email address (must be unique)","example":"user@email.com"},"password":{"type":"string","description":"User password. Must contain uppercase, lowercase and a number or special character","example":"Password123!","minLength":6,"maxLength":50,"writeOnly":true},"fullName":{"type":"string","description":"User full name","example":"John Doe","minLength":1}},"required":["email","password","fullName"]},"LoginUserDto":{"type":"object","properties":{"email":{"type":"string","description":"User email address","example":"user@email.com"},"password":{"type":"string","description":"User password. Must contain uppercase, lowercase and a number or special character","example":"Password123!","minLength":6,"maxLength":50,"writeOnly":true}},"required":["email","password"]}}}}